Make helpers and shared examples reusable
